server: created new route for passing schema validation data(maxlength) to client. client: reusable code for input handler, submit form. created input / button validation, added status messages.

+++ 03-task-manager/starter/client/src/components/Task.jsx
@@ -1,5 +1,5 @@
 //display task entry
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import './Task.css'
 import Edit from './../assets/svg_edit_2.svg';
 import Bin from './../assets/svg_bin_1.svg';
@@ -8,23 +8,18 @@ import axios from 'axios';
 import { useFormDataContext } from '../contexts/FormDataContext';
 
 export default function Task( {task, taskID} ) {
-
   const { setIsSubmittingForm } = useFormDataContext();
   const [isModalDisplayed, setIsModalDisplayed] = useState(false); 
-
 
   //hanlders
   const toggleModalHandler = () => {
     setIsModalDisplayed(prev => !prev)
   }
   const deleteTaskEntryHandler = async (id) => {
-    console.log('deleting entry with id:', id);
+    if(!id) return; 
+    setIsSubmittingForm(true);
     try {
-      const deleteRequest = await axios.delete(`/api/v1/tasks/${id}`);
-      if(deleteRequest) { 
-        setIsSubmittingForm(true);
-      }
-      return deleteRequest;
+      await axios.delete(`/api/v1/tasks/${id}`);
     } catch (error) {
       console.log(error);
     } 
@@ -34,30 +29,30 @@ export default function Task( {task, taskID} ) {
   return (
     <>
       <div className='TaskEntry'> 
-        <div className='Task' style={task.completed ? {backgroundColor: '#7c0fea'} : null}  > <p style={task.completed ? {textDecoration: 'line-through'} : null}> {task.name} </p> </div>
+        <div className='Task' style={ task.completed ? { backgroundColor: '#7c0fea' } : null }  > <p style={task.completed ? {textDecoration: 'line-through'} : null}> {task.name} </p> </div>
         <div className='TaskControl'> 
           {/* edit / delete buttons */}
           <div 
             className='ControlElem'
             onClick={toggleModalHandler}
           >
-            <img src={Edit} alt='edit button'/>
+            <img src={ Edit } alt='edit button'/>
           </div>
           <div
             onClick={async () => deleteTaskEntryHandler(task._id)} 
             className='ControlElem'
           > 
-            <img src={Bin} alt='delete button'/>
+            <img src={ Bin } alt='delete button'/>
           </div>
         </div>
       </div>
       {/* EditTask modal */}
-      {isModalDisplayed ? 
+      { isModalDisplayed ? 
       <EditTask 
-        task={task}
-        taskID={taskID}
-        toggleModalHandler={toggleModalHandler}
-      /> : null}
+        task= { task }
+        taskID={ taskID }
+        toggleModalHandler={ toggleModalHandler }
+      /> : null }
     </>
   )
 }
